app/requirements.txt
django==3.0.5
python-decouple==3.3
bpython==0.19
djangorestframework==3.11.0
psycopg2-binary==2.8.5
djangorestframework-jwt==1.11.0